+++ chutney/ui/src/app/modules/campaign/components/execution/detail/campaign-execution.component.html
@@ -1,3 +1,9 @@
+<div *ngIf="errors.length > 0" class="alert alert-dismissible alert-danger mt-4">
+    <button type="button" class="btn-close" data-bs-dismiss="alert" (click)="errors = []"></button>
+    <ng-container *ngFor="let err of errors">
+        {{ err }}<br />
+    </ng-container>
+</div>
 <div *ngIf="report">
 
     <div class="row">
